Add itineraryTab if on itinerary page
