+++ app/page/ItineraryPage.js
@@ -1,135 +1,139 @@
 import React from 'react';
 import Relay from 'react-relay';
-import { intlShape } from 'react-intl';
-import isEqual from 'lodash/isEqual';
 
-import { otpToLocation } from '../util/otp-strings';
-import config from '../config';
-import ItineraryPlanContainer from '../component/itinerary/ItineraryPlanContainer';
-import PlanRoute from '../route/PlanRoute';
-import { storeEndpoint } from '../action/EndpointActions';
-import NoRoutePopup from '../component/summary/no-route-popup';
+import CityBikeMarker from '../component/map/non-tile-layer/CityBikeMarker';
 
-const getItinerarySearch = (store) => ({
-  modes: store.getMode().join(','),
-  walkReluctance: store.getWalkReluctance(),
-  walkBoardCost: store.getWalkBoardCost(),
-  minTransferTime: store.getMinTransferTime(),
-  walkSpeed: store.getWalkSpeed(),
-  wheelchair: store.isWheelchair(),
-  maxWalkDistance: store.getMode().indexOf('BICYCLE') === -1 ?
-      config.maxWalkDistance : config.maxBikingDistance,
-  disableRemainingWeightHeuristic: store.getCitybikeState(),
-});
+import LocationMarker from '../component/map/LocationMarker';
+import ItineraryLine from '../component/map/ItineraryLine';
+import ItineraryTab from '../component/itinerary/ItineraryTab';
+import Map from '../component/map/Map';
 
-const getTime = (store) => ({
-  selectedTime: store.getSelectedTime(),
-  arriveBy: store.getArriveBy(),
-});
-
-class ItineraryPage extends React.Component {
-  static propTypes = {
-    params: React.PropTypes.shape({
-      from: React.PropTypes.string.isRequired,
-      to: React.PropTypes.string.isRequired,
-      hash: React.PropTypes.string.isRequired,
-    }).isRequired,
-  }
-
-  static contextTypes = {
-    getStore: React.PropTypes.func.isRequired,
-    intl: intlShape.isRequired,
-    router: React.PropTypes.object.isRequired,
-    raven: React.PropTypes.object.isRequired,
-  };
-
-  static loadAction(params) {
-    return [[storeEndpoint, {
-      target: 'origin',
-      endpoint: otpToLocation(params.from),
-    }], [storeEndpoint, {
-      target: 'destination',
-      endpoint: otpToLocation(params.to),
-    }]];
-  }
-
-  state = {
-    search: getItinerarySearch(this.context.getStore('ItinerarySearchStore')),
-    time: getTime(this.context.getStore('TimeStore')),
-  }
-
-  componentDidMount() {
-    this.context.getStore('ItinerarySearchStore').addChangeListener(this.onChange);
-  }
-
-  shouldComponentUpdate(newProps, newState) {
-    return !(this.state && isEqual(this.props, newProps && isEqual(this.state, newState)));
-  }
-
-  componentWillUnmount() {
-    this.context.getStore('ItinerarySearchStore').removeChangeListener(this.onChange);
-  }
-
-  onChange = () => {
-    this.setState({
-      search: getItinerarySearch(this.context.getStore('ItinerarySearchStore')),
-    });
-  }
-
-  render() {
-    // dependencies from config
-    const preferredAgencies = config.preferredAgency || '';
-
-    // dependencies from route params
-    const from = otpToLocation(this.props.params.from);
-    const to = otpToLocation(this.props.params.to);
-
-    // dependencies from itinerary search store
-    const search = this.state && this.state.search;
-
-    // dependencies from time store
-    const time = this.state && this.state.time;
-
-    // TODO: add back helmet
-
-    return (
-      <Relay.RootContainer
-        Component={ItineraryPlanContainer}
-        route={new PlanRoute({
-          fromPlace: this.props.params.from,
-          toPlace: this.props.params.to,
-          from,
-          to,
-          numItineraries: 3,
-          modes: search.modes,
-          date: time.selectedTime.format('YYYY-MM-DD'),
-          time: time.selectedTime.format('HH:mm:ss'),
-          walkReluctance: search.walkReluctance + 0.000099,
-          walkBoardCost: search.walkBoardCost,
-          minTransferTime: search.minTransferTime,
-          walkSpeed: search.walkSpeed + 0.000099,
-          maxWalkDistance: search.maxWalkDistance + 0.1,
-          wheelchair: search.wheelchair,
-
-          preferred: {
-            agencies: preferredAgencies,
-          },
-
-          arriveBy: time.arriveBy,
-          disableRemainingWeightHeuristic: search.disableRemainingWeightHeuristic,
-          hash: this.props.params.hash,
-        })}
-        renderFailure={error => {
-          this.context.raven.captureMessage('OTP returned an error when requesting a plan', {
-            extra: error,
-          });
-
-          return <div><NoRoutePopup /></div>;
-        }}
-        renderLoading={() => <div className="spinner-loader" />}
-      />
-    );
-  }
+function ItineraryPage({ itinerary }) {
+  return (
+    <ItineraryTab
+      focus={() => {}}
+      itinerary={itinerary}
+    />
+  );
 }
 
-export default ItineraryPlanContainer;
+ItineraryPage.propTypes = {
+  itinerary: React.PropTypes.object.isRequired,
+};
+
+const ItineraryPageContainer = Relay.createContainer(ItineraryPage, {
+  fragments: {
+    itinerary: () => Relay.QL`
+      fragment on Itinerary {
+        walkDistance
+        duration
+        startTime
+        endTime
+        fares {
+          type
+          currency
+          cents
+        }
+        legs {
+          mode
+          agency {
+            name
+          }
+          from {
+            lat
+            lon
+            name
+            vertexType
+            bikeRentalStation {
+              ${CityBikeMarker.getFragment('station')}
+            }
+            stop {
+              gtfsId
+              code
+              platformCode
+            }
+          }
+          to {
+            lat
+            lon
+            name
+            vertexType
+            bikeRentalStation {
+              ${CityBikeMarker.getFragment('station')}
+            }
+            stop {
+              gtfsId
+              code
+              platformCode
+            }
+          }
+          legGeometry {
+            length
+            points
+          }
+          intermediateStops {
+            gtfsId
+            lat
+            lon
+            name
+            code
+            platformCode
+          }
+          realTime
+          transitLeg
+          rentedBike
+          startTime
+          endTime
+          mode
+          distance
+          duration
+          route {
+            shortName
+            gtfsId
+          }
+          trip {
+            gtfsId
+            tripHeadsign
+            pattern {
+              code
+            }
+          }
+        }
+      }
+    `,
+  },
+});
+
+ItineraryPageContainer.renderMap = (itinerary) => {
+  const leafletObjs = [
+    <LocationMarker
+      key="from"
+      position={itinerary.legs[0].from}
+      className="from"
+    />,
+    <LocationMarker
+      key="to"
+      position={itinerary.legs[itinerary.legs.length - 1].to}
+      className="to"
+    />,
+    <ItineraryLine
+      key={'line'}
+      legs={itinerary.legs}
+      showTransferLabels
+      showIntermediateStops
+    />];
+
+  return (
+    <Map
+      className="full"
+      leafletObjs={leafletObjs}
+      lat={itinerary.legs[0].from.lat}
+      lon={itinerary.legs[0].from.lon}
+      zoom={16}
+      fitBounds={false}
+      disableZoom={false}
+    />
+  );
+};
+
+export default ItineraryPageContainer;
